+++ MDANSE/Src/MDANSE/Framework/QVectors/MillerIndicesQVectors.py
@@ -74,7 +74,7 @@ class MillerIndicesQVectors(LatticeQVectors):
 
         # The hkl matrix (3,n_hkls)
         hkls = np.mgrid[hSlice, kSlice, lSlice]
-        hkls = hkls.reshape(3, int(round(hkls.size / 3)))
+        hkls = hkls.reshape(3, round(hkls.size / 3))
 
         # The k matrix (3,n_hkls)
         vects = self.hkl_to_qvectors(hkls, self._unit_cell)
